pyproject.toml
[project]
name = "kaggle-explorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kaggle>=1.7.4.5",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
]

[project.scripts]
kaggle-explorer = "cli:main"